+++ widgetbook/lib/widgetbook/widgetbook.directories.g.dart
@@ -12,10 +12,11 @@
 import 'package:widgetbook/widgetbook.dart' as _i1;
 import 'package:widgetbook_app/widgetbook/use_cases/buttons.dart' as _i2;
 import 'package:widgetbook_app/widgetbook/use_cases/checkboxes.dart' as _i3;
-import 'package:widgetbook_app/widgetbook/use_cases/containers.dart' as _i7;
-import 'package:widgetbook_app/widgetbook/use_cases/dropdown_menus.dart' as _i4;
-import 'package:widgetbook_app/widgetbook/use_cases/running_texts.dart' as _i5;
-import 'package:widgetbook_app/widgetbook/use_cases/typography.dart' as _i6;
+import 'package:widgetbook_app/widgetbook/use_cases/containers.dart' as _i8;
+import 'package:widgetbook_app/widgetbook/use_cases/dpads.dart' as _i4;
+import 'package:widgetbook_app/widgetbook/use_cases/dropdown_menus.dart' as _i5;
+import 'package:widgetbook_app/widgetbook/use_cases/running_texts.dart' as _i6;
+import 'package:widgetbook_app/widgetbook/use_cases/typography.dart' as _i7;
 
 final directories = <_i1.WidgetbookNode>[
   _i1.WidgetbookFolder(
@@ -54,15 +55,28 @@ final directories = <_i1.WidgetbookNode>[
         ),
       ),
       _i1.WidgetbookComponent(
-        name: 'NesDropdownMenu',
+        name: 'NesDpad',
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'default',
             builder: _i4.normal,
           ),
           _i1.WidgetbookUseCase(
+            name: 'with callbacks',
+            builder: _i4.withCallbacks,
+          ),
+        ],
+      ),
+      _i1.WidgetbookComponent(
+        name: 'NesDropdownMenu',
+        useCases: [
+          _i1.WidgetbookUseCase(
+            name: 'default',
+            builder: _i5.normal,
+          ),
+          _i1.WidgetbookUseCase(
             name: 'with icons',
-            builder: _i4.icons,
+            builder: _i5.icons,
           ),
         ],
       ),
@@ -70,14 +84,14 @@ final directories = <_i1.WidgetbookNode>[
         name: 'NesRunningText',
         useCase: _i1.WidgetbookUseCase(
           name: 'default',
-          builder: _i5.normal,
+          builder: _i6.normal,
         ),
       ),
       _i1.WidgetbookLeafComponent(
         name: 'NesRunningTextLines',
         useCase: _i1.WidgetbookUseCase(
           name: 'default',
-          builder: _i5.lines,
+          builder: _i6.lines,
         ),
       ),
       _i1.WidgetbookComponent(
@@ -85,23 +99,23 @@ final directories = <_i1.WidgetbookNode>[
         useCases: [
           _i1.WidgetbookUseCase(
             name: 'body',
-            builder: _i6.body,
+            builder: _i7.body,
           ),
           _i1.WidgetbookUseCase(
             name: 'display',
-            builder: _i6.display,
+            builder: _i7.display,
           ),
           _i1.WidgetbookUseCase(
             name: 'headline',
-            builder: _i6.headline,
+            builder: _i7.headline,
           ),
           _i1.WidgetbookUseCase(
             name: 'label',
-            builder: _i6.label,
+            builder: _i7.label,
           ),
           _i1.WidgetbookUseCase(
             name: 'title',
-            builder: _i6.title,
+            builder: _i7.title,
           ),
         ],
       ),
@@ -113,19 +127,19 @@ final directories = <_i1.WidgetbookNode>[
             useCases: [
               _i1.WidgetbookUseCase(
                 name: 'default',
-                builder: _i7.normal,
+                builder: _i8.normal,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with corner inner square painter',
-                builder: _i7.cornerInnerSquare,
+                builder: _i8.cornerInnerSquare,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with label',
-                builder: _i7.label,
+                builder: _i8.label,
               ),
               _i1.WidgetbookUseCase(
                 name: 'with square corner painter',
-                builder: _i7.squareConer,
+                builder: _i8.squareConer,
               ),
             ],
           )